requirements.txt
requests>=2.2.1
python-magic>=0.4.10
